Add files via upload
The goal for Rose is to help people struggling with writer’s block by encouraging the user to generate large quantities of writing by motivating the user with graphics that grow with their character count. The writer submits a character count they wish to complete  and every time a character is entered, grass grows, and when milestones in the character count is reached, a flower will grow.